separate DecredLoading
